resolve error eslint

+++ src/controllers/products.js
@@ -1,18 +1,22 @@
 /* eslint-disable no-console */
-/* eslint-disable no-shadow */
+/* eslint-disable no-sequences */
+/* eslint-disable no-unused-expressions */
 /* eslint-disable no-plusplus */
-/* eslint-disable eqeqeq */
-const productModel = require("../models/products");
-const { v4: uuid } = require("uuid");
-const path = require("path");
-const redis = require('redis')
+/* eslint-disable consistent-return */
+/* eslint-disable array-callback-return */
+/* eslint-disable import/order */
+const productModel = require('../models/products');
+const { v4: uuid } = require('uuid');
+const path = require('path');
+const redis = require('redis');
+
 const client = redis.createClient(6379);
-const fs = require("fs/promises");
+const fs = require('fs/promises');
 
 // Handle upload image
 const uploadImageHandler = async (req) => {
   if (req.files === null) {
-    throw new Error("Image product cannot be null!");
+    throw new Error('Image product cannot be null!');
   }
 
   const files = req.files.image;
@@ -22,16 +26,16 @@ const uploadImageHandler = async (req) => {
   if (Array.isArray(files)) {
     files.map((file) => {
       if (file.size > 2 * 1024 * 1024) {
-        throw new Error("File size too large!");
+        throw new Error('File size too large!');
       }
 
-      const allowedExtension = [".png", ".jpg", ".jpeg"];
+      const allowedExtension = ['.png', '.jpg', '.jpeg'];
       const extension = path.extname(file.name);
       const fileName = `${uuid()}${extension}`;
       const outputPath = path.join(__dirname, `/../assets/images/${fileName}`);
 
       if (!allowedExtension.includes(extension)) {
-        throw new Error(`File type ${extension} are not supported!`)
+        throw new Error(`File type ${extension} are not supported!`);
       }
 
       productImages.push(fileName);
@@ -40,17 +44,17 @@ const uploadImageHandler = async (req) => {
     });
   }
 
-  if (Array.isArray(files) == false) {
-    const allowedExtension = [".png", ".jpg", ".jpeg"];
+  if (Array.isArray(files) === false) {
+    const allowedExtension = ['.png', '.jpg', '.jpeg'];
     const { image: file } = req.files;
     const extension = path.extname(file.name);
 
     if (file.size > 2 * 1024 * 1024) {
-      throw new Error("File size too large!");
+      throw new Error('File size too large!');
     }
 
     if (!allowedExtension.includes(extension)) {
-      throw new Error(`File type ${extension} are not supported!`)
+      throw new Error(`File type ${extension} are not supported!`);
     }
 
     const fileName = `${uuid()}${extension}`;
@@ -62,7 +66,7 @@ const uploadImageHandler = async (req) => {
   }
 
   return {
-    message: "Successfully uploaded",
+    message: 'Successfully uploaded',
     file_name: productImages,
   };
 };
@@ -70,27 +74,23 @@ const uploadImageHandler = async (req) => {
 // Create data to products table
 const createProduct = async (req, res, next) => {
   try {
-    if (req.user.role != 1)
+    if (req.user.role !== 1) {
       return res
         .status(400)
-        .send({ message: "you do not have access rights to add product data" });
+        .send({ message: 'you do not have access rights to add product data' });
+    }
 
-    const { title, description, categoryId, price, stock, type, color } =
-      req.body;
+    const {
+      title, description, categoryId, price, stock, type, color,
+    } = req.body;
 
-    if (!title)
-      return res.status(400).send({ message: "Title cannot be null" });
-    if (!description)
-      return res.status(400).send({ message: "Description cannot be null" });
-    if (!categoryId)
-      return res.status(400).send({ message: "Category id cannot be null" });
-    if (!price)
-      return res.status(400).send({ message: "Price cannot be null" });
-    if (!stock)
-      return res.status(400).send({ message: "Stock cannot be null" });
-    if (!type) return res.status(400).send({ message: "Type cannot be null" });
-    if (!color)
-      return res.status(400).send({ message: "Color cannot be null" });
+    if (!title) return res.status(400).send({ message: 'Title cannot be null' });
+    if (!price) return res.status(400).send({ message: 'Price cannot be null' });
+    if (!stock) return res.status(400).send({ message: 'Stock cannot be null' });
+    if (!color) return res.status(400).send({ message: 'Color cannot be null' });
+    if (!categoryId) return res.status(400).send({ message: 'Category id cannot be null' });
+    if (!type) return res.status(400).send({ message: 'Type cannot be null' });
+    if (!description) return res.status(400).send({ message: 'Description cannot be null' });
 
     const image = await uploadImageHandler(req);
 
@@ -108,7 +108,7 @@ const createProduct = async (req, res, next) => {
     await productModel.createProduct(data);
 
     res.status(201).send({
-      message: "created new category",
+      message: 'created new category',
       data,
     });
   } catch (error) {
@@ -121,22 +121,22 @@ const getProducts = (req, res, next) => {
   const { perPage } = req.query;
   const page = req.query.page || 1;
 
-  const order = req.query.orderBy || "title";
-  const sort = req.query.sortBy || "ASC";
-  const search = req.query.search || "";
+  const order = req.query.orderBy || 'title';
+  const sort = req.query.sortBy || 'ASC';
+  const search = req.query.search || '';
 
   const limit = perPage || 15;
   const offset = (page - 1) * limit;
 
   productModel
     .getAllProduct(search)
-    .then((result) => {
-      const allData = result.length;
+    .then((results) => {
+      const allData = results.length;
 
       // set cache redis all product
-      if (Object.values(req.query).length == 0) {
-        client.setex("allProduct", 60 * 60, JSON.stringify(result));
-      }
+      // if (Object.values(req.query).length === 0) {
+      //   client.setex('allProduct', 60 * 60, JSON.stringify(results));
+      // }
 
       const totalPage = Math.ceil(allData / limit);
       productModel
@@ -145,14 +145,14 @@ const getProducts = (req, res, next) => {
           if (result.length) {
             const products = result;
 
-            const resProducts = []
+            const resProducts = [];
 
             for (let i = 0; i < products.length; i++) {
-              let product = products[i]
-              const parse = JSON.parse(products[i].image)
-              product.image = parse
+              const product = products[i];
+              const parse = JSON.parse(products[i].image);
+              product.image = parse;
 
-              resProducts.push(product)
+              resProducts.push(product);
             }
 
             res.status(200);
@@ -166,18 +166,16 @@ const getProducts = (req, res, next) => {
           } else {
             res.status(404);
             res.json({
-              message: "Data not found",
+              message: 'Data not found',
             });
           }
         })
         .catch((error) => {
-          console.log(error);
-          next(new Error("Internal server error"));
+          next(new Error(error.message));
         });
     })
     .catch((error) => {
-      console.log(error);
-      next(new Error("Internal server error"));
+      next(new Error(error.message));
     });
 };
 
@@ -201,18 +199,18 @@ const getProduct = (req, res, next) => {
       });
     })
     .catch((error) => {
-      console.log(error);
-      next(new Error("Internal server error"));
+      next(new Error(error.message));
     });
 };
 
 // Update data from products table
 const updateProduct = async (req, res, next) => {
   try {
-    if (req.user.role != 1)
+    if (req.user.role !== 1) {
       return res
         .status(400)
-        .send({ message: "you do not have access rights to update product data" });
+        .send({ message: 'you do not have access rights to update product data' });
+    }
 
     const { id } = req.params;
     const {
@@ -226,24 +224,18 @@ const updateProduct = async (req, res, next) => {
       status,
     } = req.body;
 
-    if (!title)
-      return res.status(400).send({ message: "Title cannot be null" });
-    if (!description)
-      return res.status(400).send({ message: "Description cannot be null" });
-    if (!categoryId)
-      return res.status(400).send({ message: "Category id cannot be null" });
-    if (!price)
-      return res.status(400).send({ message: "Price cannot be null" });
-    if (!stock)
-      return res.status(400).send({ message: "Stock cannot be null" });
-    if (!type)
-      return res.status(400).send({ message: "Type cannot be null" });
-    if (!color)
-      return res.status(400).send({ message: "Color cannot be null" });
-    if (!status)
-      return res.status(400).send({ message: "Status cannot be null" });
+    if (!title) return res.status(400).send({ message: 'Title cannot be null' });
+    if (!description) return res.status(400).send({ message: 'Description cannot be null' });
+    if (!categoryId) return res.status(400).send({ message: 'Category id cannot be null' });
+    if (!price) return res.status(400).send({ message: 'Price cannot be null' });
+    if (!stock) return res.status(400).send({ message: 'Stock cannot be null' });
+    if (!type) return res.status(400).send({ message: 'Type cannot be null' });
+    if (!color) return res.status(400).send({ message: 'Color cannot be null' });
+    if (!status) return res.status(400).send({ message: 'Status cannot be null' });
 
-    const image = await uploadImageHandler(req);
+    let image = '';
+    if (req.files) { image = await (await uploadImageHandler(req)).file_name; }
+    if (!req.files) { image = req.body.image.split(','); }
 
     const data = {
       title,
@@ -254,7 +246,7 @@ const updateProduct = async (req, res, next) => {
       type,
       color,
       status,
-      image: JSON.stringify(image.file_name),
+      image: JSON.stringify(image),
     };
 
     const images = await productModel.getProduct(id);
@@ -262,17 +254,19 @@ const updateProduct = async (req, res, next) => {
 
     await productModel.updateProduct(data, id);
 
-    oldImages.map((img) => {
-      fs.unlink(path.join(__dirname, `/../assets/images/${img}`)),
+    if (req.files != null) {
+      oldImages.map((img) => {
+        fs.unlink(path.join(__dirname, `/../assets/images/${img}`)),
         (err) => {
           if (err) {
-            console.log("Error unlink image product!" + err);
+            console.log(`Error unlink image product!${err}`);
           }
         };
-    });
+      });
+    }
 
     res.status(200).send({
-      message: "Successfully update product!",
+      message: 'Successfully update product!',
       data,
     });
   } catch (error) {
@@ -283,54 +277,54 @@ const updateProduct = async (req, res, next) => {
 // Delete data from products table
 const deleteProduct = async (req, res, next) => {
   try {
-    if (req.user.role != 1)
+    if (req.user.role !== 1) {
       return res
         .status(400)
-        .send({ message: "you do not have access rights to delete product data" });
+        .send({ message: 'you do not have access rights to delete product data' });
+    }
 
-    const { id } = req.params
+    const { id } = req.params;
 
     const images = await productModel.getProduct(id);
     const oldImages = JSON.parse(images[0].image);
 
-    await productModel.deleteProduct(id)
+    await productModel.deleteProduct(id);
 
     oldImages.map((img) => {
       fs.unlink(path.join(__dirname, `/../assets/images/${img}`)),
-        (err) => {
-          if (err) {
-            console.log("Error unlink image product!" + err);
-          }
-        };
+      (err) => {
+        if (err) {
+          console.log(`Error unlink image product!${err}`);
+        }
+      };
     });
 
-    res.status(202)
+    res.status(202);
     res.json({
-      message: "Product successfully deleted",
-    })
+      message: 'Product successfully deleted',
+    });
   } catch (error) {
-    next(new Error("Internal server error"))
+    next(new Error('Internal server error'));
   }
 };
 
 // Get product where category
 const getProductWhereCategory = (req, res, next) => {
   const categoryId = Number(req.params.id);
-  console.log(categoryId, typeof categoryId);
   productModel
     .getProductWhereCategory(categoryId)
     .then((result) => {
       if (result.length) {
         const products = result;
 
-        const resProducts = []
+        const resProducts = [];
 
         for (let i = 0; i < products.length; i++) {
-          let product = products[i]
-          const parse = JSON.parse(products[i].image)
-          product.image = parse
+          const product = products[i];
+          const parse = JSON.parse(products[i].image);
+          product.image = parse;
 
-          resProducts.push(product)
+          resProducts.push(product);
         }
 
         res.status(200);
@@ -340,13 +334,12 @@ const getProductWhereCategory = (req, res, next) => {
       } else {
         res.status(404);
         res.json({
-          message: "Page not found",
+          message: 'Page not found',
         });
       }
     })
     .catch((error) => {
-      console.log(error);
-      next(new Error("Internal server error"));
+      next(new Error(error.message));
     });
 };
